Added calculated field on estate property form for total area.
Total area field is a calculated field, the sum of floor area
and garden area. Garden are is 0 (zero) in care the property
does not have a garden.
Also made offers visible on estate property form.

It would be nice later on to have garden area hidden when the
property has no garden. At least, have the value to be zero.